+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Completed 18.StackImplementation
 - Cunit Test framework
 - Sample Cunit test
+- Stack Tests
 
 ### Changed